Прокинул логику работы с тестированием, настроил получение данных с сервера, сверстал экраны показывающие результат тестирвоания, прокинул моковые данные

+++ data/src/main/java/com/fx_trading/lessons/data/api/question_group/ApiAnswer.kt
@@ -2,7 +2,7 @@ package com.fx_trading.lessons.data.api.question_group
 
 data class ApiAnswer(
     val is_active: Boolean,
-    val is_correct: Int,
+    val is_correct: Long,
     val optional_image_url: String,
     val text: String
 )
