+++ data/src/main/java/com/fx_trading/lessons/data/api/question_group/ApiQuestion.kt
@@ -3,7 +3,7 @@ package com.fx_trading.lessons.data.api.question_group
 data class ApiQuestion(
     val apiAnswers: List<ApiAnswer>,
     val description: String,
-    val difficulty: Int,
+    val difficulty: Long,
     val is_active: Boolean,
     val optional_image_url: String,
     val title: String
