+++ data/src/main/java/com/fx_trading/lessons/data/api/question_group/ApiQuestionGroup.kt
@@ -1,13 +1,13 @@
 package com.fx_trading.lessons.data.api.question_group
 
 data class ApiQuestionGroup(
-    val available_attempts: Int,
-    val correct_for_success: Int,
-    val id: Int,
+    val available_attempts: Long,
+    val correct_for_success: Long,
+    val id: Long,
     val is_active: Boolean,
     val language: String,
     val name: String,
-    val parent_id: Int,
+    val parent_id: Long,
     val apiQuestions: List<ApiQuestion>,
     val region: String,
     val isStartExam: Boolean
